+++ index.ts
@@ -842,3 +842,7 @@ export {
 export { getTileLoadFunction } from './src/layer/openlayers/loadFunctionHelpers.js';
 export { default as ModelFill } from './src/style/modelFill.js';
 export { renderTemplate } from './src/util/vcsTemplate.js';
+export {
+  startRotation,
+  rotationMapControlSymbol,
+} from './src/util/rotation.js';